LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/table_size64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "table_size64.wast:1"
PASS #3 Reinitialize the default imports
FAIL #4 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 47: Memory64 is not enabled at {loc} expected true got false
FAIL #5 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #6 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:10:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #7 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:13:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #8 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:16:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #9 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:19:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #10 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:22:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #11 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:25:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #12 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #13 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #14 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:34:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #15 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:37:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #16 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:40:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #17 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:43:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #18 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:46:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #19 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:49:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #20 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:52:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #21 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:55:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #22 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:58:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #23 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #24 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:64:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #25 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:67:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #26 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:70:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #27 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:73:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #28 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:76:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #29 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:79:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #30 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:82:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #31 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:85:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #32 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:88:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #33 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:91:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #34 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:94:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #35 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:97:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #36 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:100:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #37 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:103:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #38 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:106:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #39 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:109:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #40 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:112:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
FAIL #41 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
table_size64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:115:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_size64.wast.js:117:3 expected true got false
